pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = "AI chatbot with FastAPI and Weaviate"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

# Core dependencies
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
openai = "^1.14.3"
httpx = "^0.27.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
pyyaml = "^6.0.1"
weaviate-client = "3.24.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.2.2"
mypy = "^1.10.0"
isort = "^5.13.2"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
